requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
gunicorn==21.2.0
chromadb==0.5.3
pydantic
python-dotenv
sqlalchemy
crewai
langchain-openai
crewai-tools
pytest
httpx